chore(apigatewayv2): move lambda and http proxy integrations to the 'integrations' module (#11339)
We now have a `@aws-cdk/aws-apigatewayv2-integrations` module that is
intended to hold all integration of HTTP APIs with other services.

In line with this, move `LambdaProxyIntegration` and
`HttpProxyIntegration` classes to this module.

BREAKING CHANGE: `LambdaProxyIntegration` and `HttpProxyIntegration`
classes have moved to the `@aws-cdk/aws-apigatewayv2-integration` module.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*
