+++ packages/@aws-cdk/aws-apigatewayv2-integrations/lib/http/http-proxy.ts
@@ -1,5 +1,11 @@
-import { HttpIntegrationType, HttpRouteIntegrationBindOptions, HttpRouteIntegrationConfig, IHttpRouteIntegration, PayloadFormatVersion } from '../integration';
-import { HttpMethod } from '../route';
+import {
+  HttpIntegrationType,
+  HttpRouteIntegrationBindOptions,
+  HttpRouteIntegrationConfig,
+  HttpMethod,
+  IHttpRouteIntegration,
+  PayloadFormatVersion,
+} from '@aws-cdk/aws-apigatewayv2';
 
 /**
  * Properties to initialize a new `HttpProxyIntegration`.
