+++ packages/@aws-cdk/aws-apigatewayv2-integrations/lib/http/index.ts
@@ -2,3 +2,5 @@ export * from './base-types';
 export * from './alb';
 export * from './nlb';
 export * from './service-discovery';
+export * from './http-proxy';
+export * from './lambda';
