+++ packages/@aws-cdk/aws-apigatewayv2-integrations/test/http/integ.http-proxy.expected.json
