+++ packages/@aws-cdk/aws-apigatewayv2/lib/http/index.ts
@@ -1,7 +1,6 @@
 export * from './api';
 export * from './route';
 export * from './integration';
-export * from './integrations';
 export * from './stage';
 export * from './api-mapping';
 export * from './vpc-link';
