--- packages/@aws-cdk/aws-apigatewayv2/lib/http/integrations/index.ts
@@ -1,2 +1,0 @@
-export * from './http';
-export * from './lambda';
